requirements.txt
fastapi
langchain
openai
jinja2
uvicorn
pydantic
nest_asyncio
pyyaml
